scripts/requirements.txt
PyGithub == 1.59.1
PyYAML == 6.0
